fix: use new version endpoint for version check
